requirements_android.txt
kivy==2.2.1
numpy==1.24.3
tensorflow-lite==2.13.0
pyjnius==1.5.0